@@ -5,32 +5,29 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "*",
+        "php": "^7.4.11 || ^8",
         "barryvdh/laravel-ide-helper": "*",
-        "fideloper/proxy": "*",
-        "laravel/framework": "7.0.*",
-        "laravel/passport": "^8.5",
-        "laravel/socialite": "^4.1",
-        "laravel/tinker": "*",
-        "laravel/ui": "2.0",
+        "fideloper/proxy": "^4.0",
+        "laravel/framework": "8.40.*",
+        "laravel/passport": "^10.0",
+        "laravel/socialite": "^5.0",
+        "laravel/tinker": "^2.0",
         "spatie/calendar-links": "*",
         "spatie/laravel-cookie-consent": "*",
-        "lcobucci/jwt": "3.3.3"
+        "lcobucci/jwt": "4.1.4"
     },
     "require-dev": {
         "filp/whoops": "*",
         "fzaninotto/faker": "*",
         "mockery/mockery": "*",
-        "nunomaduro/collision": "*",
-        "phpunit/phpunit": "*"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
