@@ -1,9 +1,11 @@
 <?php
+namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
-use App\Models\Group;
+use Illuminate\Support\Facades\DB;
+use App\Models\GroupUser;
 
-class GroupSeeder extends Seeder
+class GroupUserSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -12,19 +14,16 @@ class GroupSeeder extends Seeder
      */
     public function run()
     {
-        Eloquent::unguard();
+        //Eloquent::unguard();
 
         // Disable Foreign key check for this connection before running seeders
         DB::statement('SET FOREIGN_KEY_CHECKS=0;');
 
-        factory(Group::class, 10)->create();
+        GroupUser::factory()
+            ->count(10)
+            ->create();
 
         DB::statement('SET FOREIGN_KEY_CHECKS=1;');
-        /*
-        Group::factory()
-            ->times(10)
-            ->create();
-        */    
     }
     
 }
