@@ -1,23 +1,32 @@
 <?php
 
+namespace Database\Factories;
+
+/** @var \Illuminate\Database\Eloquent\Factory $factory */
 use App\Models\Group;
-use Illuminate\Support\Str;
-use Faker\Generator as Faker;
+use App\Models\User;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Group::class, function (Faker $faker) {
-    return [
-        'name' => $faker->name,
-        'admin_id' => App\Models\User::first()
-    ];
-});
+class GroupFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Group::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->name,
+            'admin_id' => User::all()->random()->id,
+        ];
+    }
+}
