@@ -1,6 +1,6 @@
 <template>
     <div  v-if="calendarData['seasonData'] != undefined">
-        <div v-if="calendarData['seasonData']['seasonDraw'] > 0">
+        <div v-if="calendarData['seasonData']['is_generated'] == 1">
             Dit seizoen is al gemaakt.
             <router-link :to="{ name: 'calendar', params: { id: id }}" class="btn btn-secondary"><i class="far fa-calendar-alt"></i></router-link>
         </div>
@@ -8,19 +8,26 @@
             <h1>Generate {{  calendarData['seasonData']['name'] }}</h1>
             <button  class="btn btn-secondary" @click.prevent="generateSeason()">Regenerate season</button>
             <button  class="btn btn-secondary" @click.prevent="saveSeason()">Save season</button>
-            <two-field-two-hour-three-teams-page :calendarData="calendarData" :userData="calendarData['generateGroupUserData']" :loggedInUser="$attrs.user"  :generate="true"></two-field-two-hour-three-teams-page>              
+            <two-field-two-hour-three-teams-page :calendarData="calendarData" :userData="calendarData['groupUserData']" :loggedInUser="$attrs.user"  :generate="true"></two-field-two-hour-three-teams-page>              
         </div>
         <div v-else-if="calendarData['seasonData']['type']  == 'SingleFieldOneHourTwoTeams'">
             <h1>Generate {{  calendarData['seasonData']['name'] }}</h1>
             <button  class="btn btn-secondary" @click.prevent="generateSeason()">Regenerate season</button>
             <button  class="btn btn-secondary" @click.prevent="saveSeason()">Save season</button>
-            <single-field-one-hour-two-teams-page :calendarData="calendarData" :userData="calendarData['generateGroupUserData']" :loggedInUser="$attrs.user"  :generate="true"></single-field-one-hour-two-teams-page>  
+            <single-field-one-hour-two-teams-page :calendarData="calendarData" :userData="calendarData['groupUserData']" :loggedInUser="$attrs.user"  :generate="true"></single-field-one-hour-two-teams-page>  
         </div>
         <div v-else-if="calendarData['seasonData']['type']  == 'TwoFieldTwoHourFourTeams'">
             <h1>Generate {{  calendarData['seasonData']['name'] }}</h1>
             <button  class="btn btn-secondary" @click.prevent="generateSeason()">Regenerate season</button>
             <button  class="btn btn-secondary" @click.prevent="saveSeason()">Save season</button>
-            <two-field-two-hour-four-teams-page :calendarData="calendarData" :userData="calendarData['generateGroupUserData']" :loggedInUser="$attrs.user"  :generate="true"></two-field-two-hour-four-teams-page>              
+            <two-field-two-hour-four-teams-page :calendarData="calendarData" :userData="calendarData['groupUserData']" :loggedInUser="$attrs.user"  :generate="true"></two-field-two-hour-four-teams-page>              
+        </div>
+        <div v-else-if="calendarData['seasonData']['type']  == 'TestGenerator'" class="testGenerator">
+            <h1>Generate aaaa {{  calendarData['seasonData']['name'] }}</h1>
+            {{  calendarData['seasonData']['seasonDraw'] }}
+            <button  class="btn btn-secondary" @click.prevent="generateSeason()">Regenerate season</button>
+            <button  class="btn btn-secondary" @click.prevent="saveSeason()">Save season test</button>
+            <test-generator-page :calendarData="calendarData" :userData="calendarData['groupUserData']" :loggedInUser="$attrs.user"  :generate="true"></test-generator-page>              
         </div>
         <div v-else>
             onbekende calendar view
@@ -34,12 +41,15 @@
     import TwoFieldTwoHourThreeTeamsPage from '../CalendarPage/twoFieldTwoHourThreeTeams.vue';
     import SingleFieldOneHourTwoTeamsPage from '../CalendarPage/singleFieldOneHourTwoTeams.vue';
     import TwoFieldTwoHourFourTeamsPage from '../CalendarPage/twoFieldTwoHourFourTeams.vue';
+    import TestGeneratorPage from '../CalendarPage/testGenerator.vue';
+    
     
     export default {
         components: {
             TwoFieldTwoHourThreeTeamsPage,
             SingleFieldOneHourTwoTeamsPage,
             TwoFieldTwoHourFourTeamsPage,
+            TestGeneratorPage
         },
 
         data () {
@@ -70,10 +80,14 @@
                 }).catch(() => {
                     console.log('handle server error from here');
                 });
-                
             },
 
             saveSeason(){
+                if(this.calendarData['seasonData']['seasonDraw'] > 0){
+                    this.saveGeneratedTeamsSeason();
+                    return;
+                }
+
                  this.formData.set('jsonSeason', JSON.stringify(this.calendarData));
                  apiCall.postData('season/' +  this.id + '/generator', this.formData)
                 .then(response =>{
@@ -83,6 +97,19 @@
                 }).catch(error => {
                     this.errors = error;
                 });                
+            },
+
+            saveGeneratedTeamsSeason(){
+                this.formData.append('teamRange', JSON.stringify(this.calendarData['data']));
+                this.formData.append('updateRange', 'pregeneratedseason');
+                apiCall.postData('season/' +  this.id + '/generator/' + this.id, this.formData)
+                .then(response =>{
+                     this.message = "Your season has been created";
+                    this.$bus.$emit('showMessage', this.message,  'green', '2000' );
+                    router.push({ name: 'calendar', params: { id:  this.id }});                   
+                }).catch(error => {
+                    this.errors = error;
+                });
             },
         },
 
