@@ -10,7 +10,11 @@ let mix = require('laravel-mix');
  | file for the application as well as bundling up all the JS files.
  |
  */
-
+/*
 mix.setPublicPath('public_html/');
 mix.js('resources/js/app.js', 'js')
     .sass('resources/sass/app.scss', 'css');
+*/
+mix.setPublicPath('public_html/');
+mix.js('resources/js/app.js', 'public_html/js').vue();
+mix.sass('resources/sass/app.scss', 'public_html/css');
